exceptions: separate exceptions module introduced
The ratatoskr engine related exceptions have been assorted
to the 'exceptions' module which is available as ratatoskr.exceptions,
resulting in the usage example:

import ratatoskr

try:
  ...
  except ratatoskr.exceptions.SchemaValidationError as e:
  ...
