+++ ratatoskr/__init__.py
@@ -2,6 +2,8 @@ import utils
 import operation_registry
 import operation_wrappers.base_wrappers as base_wrappers
 import types
+import exceptions
+
 from protectron import protectron
