+++ ratatoskr/operation_registry.py
@@ -2,18 +2,11 @@ from protectron import protectron
 from schema import ValidOperationRegistryEventSchema
 from operation_wrappers.base_wrappers import OperationWrapper
 from internal_logger import LOG
-
-
-class InvalidOperationWrapperError(Exception):
-    pass
-
-
-class OperationAlreadyRegisteredError(Exception):
-    pass
-
-
-class UnregisteredOperationError(Exception):
-    pass
+from exceptions import (
+    InvalidOperationWrapperError,
+    OperationAlreadyRegisteredError,
+    UnregisteredOperationError
+)
 
 
 class OperationRegistry:
